// crm_frontend_3/src/api/api.ts
export interface Create_Order_ClientDto {
    name_Client: string ;
    lastName_Client: string ;
    email_Client: string;
    telefon: string ;
    type_technology: string ;
    model_technology: string ;
    breaking_info: string;
    iD_Personnel_master: string;
}
// ...
export interface List_Personnal {
    iD_Personnal: string;
    name: string;
    last_Name: string;
    telefon: string;
    position: string;
}